Voting3 V2.0 大修版本
改动了很多数据类型，int->double, float->double ，之前由于用Int来存double的结果，导致结果有误差

+++ Voting3/LP.h
@@ -71,7 +71,7 @@ double LP(PGraph *g,vector<Req*> &reqL)
 
 		//展示流量所走的路径，并且修改各条link的capacity
 		
-		float latency=0;
+		double latency=0;
 		for(int d=0;d<K;d++)
 		{
 			//cout<<"flow "<<d+1<<" : "<<endl;
